requirements.txt
lxml
roman
python-slugify[unidecode]
sphinx
